pyproject.toml
[project]
name = "mlflow-oss-databricks-apps"
version = "0.1.0"
description = "MLflow Open Source Server on Databricks Apps"
authors = [
    {name = "Nikhil Thorat", email = "nikhil.thorat@databricks.com"},
]
dependencies = [
    "mlflow>=2.19.0",
    "databricks-sdk>=0.33.0",
    "gunicorn>=22.0.0",
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.ruff]
line-length = 120
fix = true

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"